requirements.txt
fastapi~=0.115.12
uvicorn[standard]
mysql-connector-python~=9.3.0

pydantic~=2.11.5